+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - feature: install hermes agent
 - feature: install insomnia
 - feature: install jetbrains toolbox
+- feature: install lm studio
 
 ## [1.4.0] - 2026-07-24